refactoring (split functions into multiple files)
